@@ -5,7 +5,7 @@
   "license"     : "GNU GPLv2",
   "type"        : "project",
   "require"     : {
-    "php": "^8.0",
+    "php": "^8.2.0",
     "alaouy/youtube": "^2.0",
     "anhskohbo/no-captcha": "^3.0",
     "bugsnag/bugsnag-laravel": "^2.0",
@@ -13,26 +13,23 @@
     "doctrine/dbal": "^3.0",
     "eluceo/ical": "^0.15",
     "eonasdan/bootstrap-datetimepicker": "^4.17",
-    "facade/ignition": "^2.3.6",
-    "fideloper/proxy": "^4.0",
     "fortawesome/font-awesome": "^4.7",
-    "fruitcake/laravel-cors": "^2.0",
-    "graham-campbell/markdown": "^13.0",
-    "guzzlehttp/guzzle": "^7.0.1",
+    "graham-campbell/markdown": "^15.2",
+    "guzzlehttp/guzzle": "^7.2",
     "intervention/image": "^2.3",
+    "konekt/html": "^6.5",
     "laminas/laminas-stdlib": "^3.0",
-    "laravel/framework": "^8.0",
-    "laravel/slack-notification-channel": "^2.0",
-    "laravel/tinker": "^2.0",
-    "laravel/ui": "^3.0",
-    "laravelcollective/html": "^6.0",
+    "laravel/framework": "^11.0",
+    "laravel/tinker": "^2.7",
+    "laravel/ui": "^4.0",
     "lavary/laravel-menu": "^1.7",
     "mews/purifier": "^3.0",
     "moment/moment": "^2.17",
-    "nunomaduro/collision": "^5.3",
+    "nunomaduro/collision": "^8.1",
     "select2/select2": "^4.0",
-    "spatie/laravel-backup": "^7.0",
-    "spatie/laravel-db-snapshots": "^1.1",
+    "spatie/laravel-backup": "^8.6",
+    "spatie/laravel-db-snapshots": "^2.6",
+    "spatie/laravel-ignition": "^2.0",
     "twbs/bootstrap": "^3.3"
   },
   "require-dev" : {
@@ -40,8 +37,8 @@
     "fakerphp/faker": "^1.9.1",
     "filp/whoops": "^2.0",
     "mockery/mockery": "^1.4.2",
-    "phpunit/php-code-coverage": "^9.2",
-    "phpunit/phpunit": "^9.5"
+    "phpunit/php-code-coverage": "^10.0",
+    "phpunit/phpunit": "^10.0"
   },
   "autoload"    : {
     "psr-4": {
@@ -77,7 +74,12 @@
   "config"      : {
     "optimize-autoloader": true,
     "preferred-install"  : "dist",
-    "sort-packages"      : true
+    "sort-packages"      : true,
+    "allow-plugins": {
+      "composer/installers": true,
+      "oomphinc/composer-installers-extender": true,
+      "robloach/component-installer": true
+    }
   },
   "repositories": [
   ]
